+++ ChangeLog.md
@@ -14,6 +14,7 @@
 - Modify ChangeLog.md etc position to ${CMAKE_INSTALL_DOCDIR}
 - RabbitCommonUtils: Add INSTALL_RPATH  
   FIX: 7 in https://github.com/KangLin/RabbitRemoteControl/issues/20
+- Update: fix CFrmUpdater::CompareVersion bug
 
 ### Version: v2.0.3
 - Fix: fix footer.html bug